feat(smoke): extend yaml-driven smoke testing to workflow tests
This commit introduces a new harness and smoke-style test runner to
simulate the sequence of plugin create, add, enable-ment, and the new
`plugin remove` commands.

The testing environment is isolated via `--coll-root` and produces
full-fat plugin archetyping from a fully validate bundled plugin. The
sequence of commands requires preservation of state across steps. This
is a new introduction to `test/runners/smoke/`, as previously the tests
here only ran non-state changing commands.

In addition to the CRUD closure of the plugin command now having a
remove method, dynamic tests and advanced workflows can now be ran with
optics. The addition of the `--show` flag to the original two smoke
tests (one formatter checking, the other execution smoking) allowed
those to be consolidated into one YAML file.

This new environment allows for better automated interactive testing.
In subsequent commits, the other Level 4 tests will be ported over to
this new system; we maintainers need real hands on these scenarios.
Quantitative breadth of inter-modularity testing means nothing if the
app doesn't qualitatively behave as you intended.

In addition, the YAML format offers other advantageous. Self
-documentation and tagging along with declarations impossible to shift
in transit means the sequence of commands and their expectations can be
in one source, together with the test's abstract, providing a central
authority. Eventually, this will see the union of:

- `test/config/metadata-level-4.yaml`           (docs)
- `test/runners/smoke/workflow-tests.yaml`      (neo-manifest)
- `test/runners/e2e/workflow-lifecycle.test.js` (crusty imperatus)