feat(test): begin porting of E2E manifests to YAML driver
This commit begins reorganizing the pilot, static End-to-End tests from
the old `*.js` based manifests to the new YAML format.  To ensure
parity, we have broken the manifests into smaller blocks to directly
compare to their `*.js` cousins.

The next commit should port the remainder of the manifests to YAML.