feat(test): port all remaining E2E manifests to YAML
This commit finishes the porting of the old JS manifests to the new YAML
manifest system, and adds new features to the YAML test runner and
workflow system, including the ability to add debug steps to workflows,
introduces better in-situ templating, and resolves a newly discovered
discrepancy found between the XDG vs. `--coll-root` usage in
`collection add / remove`

The main points.

- No substantial code generated.

  The newly designed system required practically no new code to
  implement the remaining 3/4 of E2E tests. Formerly, this would have
  required significant debugging of JS and NOT implementation of of
  checks for plugin integrity.

- Performance of old vs. new system.

  - Old System:  2m41.349s
    `time npm test -- --group e2e`
  - New System:  2m43.915s
    `time node test/runners/smoke/yaml-test-runner.js \
          test/runners/smoke/*.yaml`

- Maintainability.

  - Old system:  1016 lines of manifest code
    `wc -l ../e2e/*.manifest.js | tail -1`
  - New System:  528 lines
    `wc -l *.yaml | tail -1`

- Coverage.

   - Old system:  38
     `grep -R "describe:" ../e2e/*.manifest.js | wc -l`
   - New system:  64
     `grep -R "description:" *.yaml | wc -l`

We have reduced the amount of boilerplate entirely, have implemented a
format that provides automatic dosumentation, reduces lines of code by a
factor of 2, increased edge case coverage, using full-fat real
plugins--not minimal mocks--with no performance loss.

The spin-up time to online testing is the most important heuristic. New
tests are as trivial as documenting a line of shell.